pyproject.toml
[tool.poetry]
name = "wordle-on-telegram"
version = "1.0.0"
description = ""
authors = ["Gennadiy Matveev <geno.matveev@gmail.com>"]
readme = "README.md"
packages = [{include = "wordle_on_telegram"}]
license = "MIT"


[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.2"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.27"}
sqladmin = "^0.16.1"
aiogram = "3.3.0"
environs = "^10.3.0"
uvicorn = "^0.27.1"
alembic = "^1.13.1"
asyncpg = "^0.29.0"
pyjwt = "^2.8.0"
itsdangerous = "^2.1.2"
gunicorn = "^21.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.ruff]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]

# Same as Black.
line-length = 120
indent-width = 4

target-version = "py38"

[tool.ruff.lint]
# Enable Pyflakes (`F`) and a subset of the pycodestyle (`E`)  codes by default.
select = ["E4", "E7", "E9", "F"]
ignore = ["F401", "E712"]

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = ["F401"]

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "single"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

docstring-code-format = true
